+++ packages/concerto-cli/demo.json
@@ -1,0 +1,1 @@
+{"$class":"concerto.metamodel.Model","namespace":"org.accordproject.cicero.contract","imports":[],"declarations":[{"$class":"concerto.metamodel.AssetDeclaration","name":"AccordContractState","isAbstract":false,"properties":[{"$class":"concerto.metamodel.StringProperty","name":"stateId","isArray":false,"isOptional":false,"location":{"$class":"concerto.metamodel.Range","start":{"offset":809,"line":24,"column":3,"$class":"concerto.metamodel.Position"},"end":{"offset":826,"line":25,"column":1,"$class":"concerto.metamodel.Position"}}}],"location":{"$class":"concerto.metamodel.Range","start":{"offset":757,"line":23,"column":1,"$class":"concerto.metamodel.Position"},"end":{"offset":827,"line":25,"column":2,"$class":"concerto.metamodel.Position"}},"identified":{"$class":"concerto.metamodel.IdentifiedBy","name":"stateId"}},{"$class":"concerto.metamodel.ParticipantDeclaration","name":"AccordParty","isAbstract":false,"properties":[{"$class":"concerto.metamodel.StringProperty","name":"partyId","isArray":false,"isOptional":false,"location":{"$class":"concerto.metamodel.Range","start":{"offset":907,"line":29,"column":3,"$class":"concerto.metamodel.Position"},"end":{"offset":924,"line":30,"column":1,"$class":"concerto.metamodel.Position"}}}],"location":{"$class":"concerto.metamodel.Range","start":{"offset":857,"line":28,"column":1,"$class":"concerto.metamodel.Position"},"end":{"offset":925,"line":30,"column":2,"$class":"concerto.metamodel.Position"}},"identified":{"$class":"concerto.metamodel.IdentifiedBy","name":"partyId"}},{"$class":"concerto.metamodel.AssetDeclaration","name":"AccordContract","isAbstract":true,"properties":[{"$class":"concerto.metamodel.StringProperty","name":"contractId","isArray":false,"isOptional":false,"location":{"$class":"concerto.metamodel.Range","start":{"offset":1049,"line":34,"column":3,"$class":"concerto.metamodel.Position"},"end":{"offset":1071,"line":35,"column":3,"$class":"concerto.metamodel.Position"}}},{"$class":"concerto.metamodel.RelationshipProperty","name":"parties","type":{"$class":"concerto.metamodel.TypeIdentifier","name":"AccordParty"},"isArray":true,"isOptional":true,"location":{"$class":"concerto.metamodel.Range","start":{"offset":1071,"line":35,"column":3,"$class":"concerto.metamodel.Position"},"end":{"offset":1106,"line":36,"column":1,"$class":"concerto.metamodel.Position"}}}],"location":{"$class":"concerto.metamodel.Range","start":{"offset":990,"line":33,"column":1,"$class":"concerto.metamodel.Position"},"end":{"offset":1107,"line":36,"column":2,"$class":"concerto.metamodel.Position"}},"identified":{"$class":"concerto.metamodel.IdentifiedBy","name":"contractId"}},{"$class":"concerto.metamodel.AssetDeclaration","name":"AccordClause","isAbstract":true,"properties":[{"$class":"concerto.metamodel.StringProperty","name":"clauseId","isArray":false,"isOptional":false,"location":{"$class":"concerto.metamodel.Range","start":{"offset":1224,"line":40,"column":3,"$class":"concerto.metamodel.Position"},"end":{"offset":1242,"line":41,"column":1,"$class":"concerto.metamodel.Position"}}}],"location":{"$class":"concerto.metamodel.Range","start":{"offset":1169,"line":39,"column":1,"$class":"concerto.metamodel.Position"},"end":{"offset":1243,"line":41,"column":2,"$class":"concerto.metamodel.Position"}},"identified":{"$class":"concerto.metamodel.IdentifiedBy","name":"clauseId"}}]}
