+++ packages/concerto-core/lib/typenotfoundexception.js
@@ -14,7 +14,7 @@
 
 'use strict';
 
-const BaseException = require('./baseexception');
+const BaseException = require('@accordproject/concerto-cto').BaseException;
 const Globalize = require('./globalize');
 
 /**
